changed structure + added factories

+++ WorkflowProcessingModel/Model/SubElement/Setup.cs
@@ -5,9 +5,9 @@ namespace WorkflowProcessingModel.Model
     abstract class Setup
     {
         public Machine CurrentMachine { get; set; }
-        public double SetupTime { get; set; }
+        public int SetupTime { get; set; }
 
-        public Setup(Machine currentMachine, double setupTime)
+        public Setup(Machine currentMachine, int setupTime)
         {
             CurrentMachine = currentMachine ?? throw new ArgumentNullException(nameof(currentMachine));
             SetupTime = setupTime;
