+++ WorkflowProcessingModel/Model/SubElement/Storehouse.cs
@@ -8,9 +8,9 @@ namespace WorkflowProcessingModel.Model
         public int Index { get; set; }
         public string Name { get; set; }
         public Dictionary<Material, int> AvailableMaterials { get; set; }
-        public double TimeNeededToTransportProductsToProductionHall { get; set; }
+        public int TimeNeededToTransportProductsToProductionHall { get; set; }
 
-        public Storehouse(int index, string name, Dictionary<Material, int> availableMaterials, double timeNeededToTransportProductsToProductionHall)
+        public Storehouse(int index, string name, Dictionary<Material, int> availableMaterials, int timeNeededToTransportProductsToProductionHall)
         {
             Index = index;
             Name = name ?? throw new ArgumentNullException(nameof(name));
